automatically delete downloaded document from cache.

+++ feature/client/src/commonMain/kotlin/com/mifos/feature/client/DocumentSelectAndUploadRepositoryImpl.kt
@@ -19,6 +19,7 @@ import com.mifos.core.data.repository.DocumentListRepository
 import com.mifos.feature.client.utils.createDocumentRequestBody
 import io.github.vinceglb.filekit.PlatformFile
 import io.github.vinceglb.filekit.div
+import io.github.vinceglb.filekit.parent
 import io.ktor.client.request.forms.MultiPartFormDataContent
 import io.ktor.client.statement.readRawBytes
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -159,7 +160,7 @@ class DocumentSelectAndUploadRepositoryImpl(
     override suspend fun deleteDocumentFormCache() {
         val state = entityDocumentStateMutableStateFlow.first()
         state.entityDocument?.let {
-            FileKitUtil.deleteFile(it)
+            if (it.parent() == FileKitUtil.appCache) FileKitUtil.deleteFile(it)
         }
     }
